Implemented controllers, services and tests demonstrating the differences between Mock, Stub and Spy.
I am checking this in for posterity sake, but will delete the `SpringSecurity` examples because it's a bit overkill for this simple example.  My initial thought was that it would be good to show Mocks, Stubs and Spys used in a real world example with `SpringSecurity`, but it complicated things and detracted from the focus of the blog post which is simpley Mock vs Stub vs Spy.

As a result, I've created  simpler examples.